patch: extract CellularSettings to its own module
